Move base API class to separate file
